edit to var name

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -6,8 +6,9 @@ from typing import List
 from typing import Union
 
 
-def sum_mixed_list(mxd_lst: List[Union[float, int]]) -> float:
+def sum_mixed_list(mxd_lst: List[Union[int, float]]) -> float:
     """
     takes mxd_lst of integers/floats and returns their sum as float.
     """
+
     return sum(mxd_lst)
